Isolate JUnit4 utility code #857
Task-Url: https://github.com/eclipse-mylyn/org.eclipse.mylyn/issues/issues/857
