+++ mylyn.builds/org.eclipse.mylyn.jenkins.tests/META-INF/MANIFEST.MF
@@ -11,6 +11,7 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="0.0.0",
  org.eclipse.mylyn.commons.core;bundle-version="4.4.0",
  org.eclipse.mylyn.commons.repositories.core;bundle-version="4.4.0",
  org.eclipse.mylyn.commons.sdk.util;bundle-version="4.4.0",
+ org.eclipse.mylyn.commons.sdk.util.junit4;bundle-version="4.8.0",
  org.eclipse.mylyn.jenkins.core;bundle-version="4.4.0",
  org.eclipse.mylyn.jenkins.ui;bundle-version="4.4.0",
  org.eclipse.search;bundle-version="0.0.0",
