+++ mylyn.builds/org.eclipse.mylyn.jenkins.tests/src/org/eclipse/mylyn/jenkins/tests/AllJenkinsTests.java
@@ -14,9 +14,9 @@ package org.eclipse.mylyn.jenkins.tests;
 import java.util.List;
 
 import org.eclipse.mylyn.commons.sdk.util.CommonTestUtil;
-import org.eclipse.mylyn.commons.sdk.util.ManagedSuite;
-import org.eclipse.mylyn.commons.sdk.util.ManagedTestSuite;
 import org.eclipse.mylyn.commons.sdk.util.TestConfiguration;
+import org.eclipse.mylyn.commons.sdk.util.junit4.ManagedSuite;
+import org.eclipse.mylyn.commons.sdk.util.junit4.ManagedTestSuite;
 import org.eclipse.mylyn.jenkins.tests.client.JenkinsClientTest;
 import org.eclipse.mylyn.jenkins.tests.client.JenkinsUrlTest;
 import org.eclipse.mylyn.jenkins.tests.client.JenkinsValidationTest;
