+++ mylyn.builds/org.eclipse.mylyn.jenkins.tests/src/org/eclipse/mylyn/jenkins/tests/support/JenkinsFixture.java
@@ -20,8 +20,8 @@ import org.eclipse.core.runtime.Status;
 import org.eclipse.mylyn.commons.core.StatusHandler;
 import org.eclipse.mylyn.commons.repositories.core.RepositoryLocation;
 import org.eclipse.mylyn.commons.sdk.util.FixtureConfiguration;
-import org.eclipse.mylyn.commons.sdk.util.RepositoryTestFixture;
 import org.eclipse.mylyn.commons.sdk.util.TestConfiguration;
+import org.eclipse.mylyn.commons.sdk.util.junit4.RepositoryTestFixture;
 import org.eclipse.mylyn.internal.jenkins.core.JenkinsCorePlugin;
 import org.eclipse.mylyn.internal.jenkins.core.client.JenkinsConfigurationCache;
 import org.eclipse.mylyn.internal.jenkins.core.client.JenkinsServerInfo.Type;
