+++ mylyn.commons/org.eclipse.mylyn.commons.sdk.util.junit4/META-INF/MANIFEST.MF
@@ -1,0 +1,14 @@
+Manifest-Version: 1.0
+Automatic-Module-Name: org.eclipse.mylyn.commons.sdk.util.junit4
+Bundle-ManifestVersion: 2
+Bundle-Name: Mylyn Commons SDK JUnit 4 Utilities
+Bundle-SymbolicName: org.eclipse.mylyn.commons.sdk.util.junit4;singleton:=true
+Bundle-Version: 4.8.0.qualifier
+Bundle-RequiredExecutionEnvironment: JavaSE-21
+Require-Bundle: org.eclipse.core.runtime;bundle-version="0.0.0",
+ org.eclipse.mylyn.commons.core;bundle-version="4.8.0",
+ org.eclipse.mylyn.commons.sdk.util;bundle-version="4.8.0",
+ org.eclipse.osgi;bundle-version="0.0.0",
+ org.junit;bundle-version="4.8.2"
+Export-Package: org.eclipse.mylyn.commons.sdk.util.junit4;x-internal:=true
+Bundle-Vendor: Eclipse Mylyn
