+++ mylyn.commons/org.eclipse.mylyn.commons.sdk.util.junit4/build.properties
@@ -1,0 +1,18 @@
+###############################################################################
+# Copyright (c) 2009, 2012 Tasktop Technologies and others.
+# 
+# This program and the accompanying materials are made available under the
+# terms of the Eclipse Public License v. 2.0 which is available at
+# https://www.eclipse.org/legal/epl-2.0
+# 
+# SPDX-License-Identifier: EPL-2.0
+#
+# Contributors:
+#      Tasktop Technologies - initial API and implementation
+###############################################################################
+source.. = src/
+output.. = bin/
+bin.includes = .,\
+               META-INF/,\
+               about.html
+src.includes = about.html
