+++ mylyn.commons/org.eclipse.mylyn.commons.sdk.util.junit4/src/org/eclipse/mylyn/commons/sdk/util/junit4/MustRunOnCIServerRule.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2016 Frank Becker and others.
+ * Copyright (c) 2024 Frank Becker and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -7,18 +7,21 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *
+ * Contributors:
  *     Frank Becker - initial API and implementation
+ *     See git history
  *******************************************************************************/
 
-package org.eclipse.mylyn.commons.sdk.util;
+package org.eclipse.mylyn.commons.sdk.util.junit4;
 
-import org.junit.runners.model.Statement;
+import org.eclipse.mylyn.commons.sdk.util.AbstractTestFixture;
+import org.eclipse.mylyn.commons.sdk.util.CommonTestUtil;
 
-public class IgnoreStatement extends Statement {
+public class MustRunOnCIServerRule implements ConditionalIgnoreRule.IgnoreCondition {
 
 	@Override
-	public void evaluate() {
-		throw new IgnoreRuleRuntimeException();
+	public boolean isSatisfied(AbstractTestFixture fixture) {
+		return !CommonTestUtil.runOnCIServerTestsOnly();
 	}
 
 }
