+++ mylyn.reviews/org.eclipse.mylyn.gerrit.tests/META-INF/MANIFEST.MF
@@ -14,6 +14,7 @@ Require-Bundle: net.bytebuddy.byte-buddy;bundle-version="1.12.18",
  org.eclipse.mylyn.commons.net;bundle-version="4.4.0",
  org.eclipse.mylyn.commons.repositories.core;bundle-version="4.4.0",
  org.eclipse.mylyn.commons.sdk.util;bundle-version="4.4.0",
+ org.eclipse.mylyn.commons.sdk.util.junit4;bundle-version="4.8.0",
  org.eclipse.mylyn.commons.ui;bundle-version="4.4.0",
  org.eclipse.mylyn.commons.workbench;bundle-version="4.4.0",
  org.eclipse.mylyn.gerrit.core;bundle-version="4.4.0",
