+++ mylyn.tasks/connectors/bugzilla-rest/org.eclipse.mylyn.bugzilla.rest.core.tests/META-INF/MANIFEST.MF
@@ -16,6 +16,7 @@ Require-Bundle: org.hamcrest;bundle-version="0.0.0",
  org.eclipse.mylyn.commons.repositories.core;bundle-version="4.4.0",
  org.eclipse.mylyn.commons.repositories.http.core;bundle-version="4.4.0",
  org.eclipse.mylyn.commons.sdk.util;bundle-version="4.4.0",
+ org.eclipse.mylyn.commons.sdk.util.junit4;bundle-version="4.8.0",
  org.eclipse.mylyn.tasks.core;bundle-version="4.4.0",
  org.junit;bundle-version="4.8.2"
 Export-Package: org.eclipse.mylyn.bugzilla.rest.core.tests;x-internal:=true,
