+++ mylyn.tasks/connectors/bugzilla-rest/org.eclipse.mylyn.bugzilla.rest.core.tests/src/org/eclipse/mylyn/bugzilla/rest/core/tests/AllBugzillaRestCoreTests.java
@@ -17,10 +17,10 @@ package org.eclipse.mylyn.bugzilla.rest.core.tests;
 import java.util.List;
 
 import org.eclipse.mylyn.commons.sdk.util.CommonTestUtil;
-import org.eclipse.mylyn.commons.sdk.util.ManagedSuite;
-import org.eclipse.mylyn.commons.sdk.util.ManagedSuite.SuiteClassProvider;
-import org.eclipse.mylyn.commons.sdk.util.ManagedSuite.TestConfigurationProperty;
 import org.eclipse.mylyn.commons.sdk.util.TestConfiguration;
+import org.eclipse.mylyn.commons.sdk.util.junit4.ManagedSuite;
+import org.eclipse.mylyn.commons.sdk.util.junit4.ManagedSuite.SuiteClassProvider;
+import org.eclipse.mylyn.commons.sdk.util.junit4.ManagedSuite.TestConfigurationProperty;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
