+++ mylyn.tasks/connectors/bugzilla-rest/org.eclipse.mylyn.bugzilla.rest.core.tests/src/org/eclipse/mylyn/bugzilla/rest/core/tests/BugzillaRestClientTest.java
@@ -49,11 +49,11 @@ import org.eclipse.mylyn.commons.repositories.core.auth.AuthenticationType;
 import org.eclipse.mylyn.commons.repositories.core.auth.UserCredentials;
 import org.eclipse.mylyn.commons.sdk.util.AbstractTestFixture;
 import org.eclipse.mylyn.commons.sdk.util.CommonTestUtil;
-import org.eclipse.mylyn.commons.sdk.util.ConditionalIgnoreRule;
-import org.eclipse.mylyn.commons.sdk.util.IFixtureJUnitClass;
-import org.eclipse.mylyn.commons.sdk.util.Junit4TestFixtureRunner;
-import org.eclipse.mylyn.commons.sdk.util.Junit4TestFixtureRunner.FixtureDefinition;
-import org.eclipse.mylyn.commons.sdk.util.MustRunOnCIServerRule;
+import org.eclipse.mylyn.commons.sdk.util.junit4.ConditionalIgnoreRule;
+import org.eclipse.mylyn.commons.sdk.util.junit4.IFixtureJUnitClass;
+import org.eclipse.mylyn.commons.sdk.util.junit4.Junit4TestFixtureRunner;
+import org.eclipse.mylyn.commons.sdk.util.junit4.Junit4TestFixtureRunner.FixtureDefinition;
+import org.eclipse.mylyn.commons.sdk.util.junit4.MustRunOnCIServerRule;
 import org.eclipse.mylyn.internal.bugzilla.rest.core.BugzillaRestAttachmentMapper;
 import org.eclipse.mylyn.internal.bugzilla.rest.core.BugzillaRestClient;
 import org.eclipse.mylyn.internal.bugzilla.rest.core.BugzillaRestConfiguration;
