+++ mylyn.tasks/connectors/bugzilla-rest/org.eclipse.mylyn.bugzilla.rest.core.tests/src/org/eclipse/mylyn/bugzilla/rest/core/tests/BugzillaRestConfigurationTest.java
@@ -25,11 +25,11 @@ import org.eclipse.core.runtime.CoreException;
 import org.eclipse.mylyn.bugzilla.rest.test.support.BugzillaRestTestFixture;
 import org.eclipse.mylyn.commons.sdk.util.AbstractTestFixture;
 import org.eclipse.mylyn.commons.sdk.util.CommonTestUtil;
-import org.eclipse.mylyn.commons.sdk.util.ConditionalIgnoreRule;
-import org.eclipse.mylyn.commons.sdk.util.IFixtureJUnitClass;
-import org.eclipse.mylyn.commons.sdk.util.Junit4TestFixtureRunner;
-import org.eclipse.mylyn.commons.sdk.util.Junit4TestFixtureRunner.FixtureDefinition;
-import org.eclipse.mylyn.commons.sdk.util.MustRunOnCIServerRule;
+import org.eclipse.mylyn.commons.sdk.util.junit4.ConditionalIgnoreRule;
+import org.eclipse.mylyn.commons.sdk.util.junit4.IFixtureJUnitClass;
+import org.eclipse.mylyn.commons.sdk.util.junit4.Junit4TestFixtureRunner;
+import org.eclipse.mylyn.commons.sdk.util.junit4.Junit4TestFixtureRunner.FixtureDefinition;
+import org.eclipse.mylyn.commons.sdk.util.junit4.MustRunOnCIServerRule;
 import org.eclipse.mylyn.internal.bugzilla.rest.core.BugzillaRestConfiguration;
 import org.eclipse.mylyn.internal.bugzilla.rest.core.BugzillaRestConnector;
 import org.eclipse.mylyn.internal.tasks.core.TaskRepositoryManager;
