+++ mylyn.tasks/connectors/bugzilla/org.eclipse.mylyn.bugzilla.tests/META-INF/MANIFEST.MF
@@ -13,6 +13,7 @@ Require-Bundle: org.eclipse.compare;bundle-version="0.0.0",
  org.eclipse.mylyn.commons.net;bundle-version="4.4.0",
  org.eclipse.mylyn.commons.repositories.core;bundle-version="4.4.0",
  org.eclipse.mylyn.commons.sdk.util;bundle-version="4.4.0",
+ org.eclipse.mylyn.commons.sdk.util.junit4;bundle-version="4.8.0",
  org.eclipse.mylyn.commons.ui;bundle-version="4.4.0",
  org.eclipse.mylyn.commons.workbench;bundle-version="4.4.0",
  org.eclipse.mylyn.commons.xmlrpc;bundle-version="4.4.0",
