+++ org.eclipse.mylyn.tests/META-INF/MANIFEST.MF
@@ -16,6 +16,7 @@ Require-Bundle: org.eclipse.core.resources;bundle-version="0.0.0",
  org.eclipse.mylyn.cdt.tests;bundle-version="4.4.0",
  org.eclipse.mylyn.commons.notifications.tests;bundle-version="4.4.0",
  org.eclipse.mylyn.commons.sdk.util;bundle-version="4.4.0",
+ org.eclipse.mylyn.commons.sdk.util.junit4;bundle-version="4.8.0",
  org.eclipse.mylyn.commons.tests;bundle-version="4.4.0",
  org.eclipse.mylyn.context.core;bundle-version="4.4.0",
  org.eclipse.mylyn.context.tasks.tests;bundle-version="4.4.0",
