+++ org.eclipse.mylyn.tests/src/org/eclipse/mylyn/tests/AllConnectorHeartbeatTests.java
@@ -12,9 +12,9 @@
 
 package org.eclipse.mylyn.tests;
 
-import org.eclipse.mylyn.commons.sdk.util.ManagedSuite;
-import org.eclipse.mylyn.commons.sdk.util.ManagedTestSuite;
 import org.eclipse.mylyn.commons.sdk.util.TestConfiguration;
+import org.eclipse.mylyn.commons.sdk.util.junit4.ManagedSuite;
+import org.eclipse.mylyn.commons.sdk.util.junit4.ManagedTestSuite;
 
 import junit.framework.JUnit4TestAdapter;
 import junit.framework.Test;
