+++ org.eclipse.mylyn.tests/src/org/eclipse/mylyn/tests/AllHeartbeatTests.java
@@ -12,8 +12,8 @@
 
 package org.eclipse.mylyn.tests;
 
-import org.eclipse.mylyn.commons.sdk.util.ManagedTestSuite;
 import org.eclipse.mylyn.commons.sdk.util.TestConfiguration;
+import org.eclipse.mylyn.commons.sdk.util.junit4.ManagedTestSuite;
 
 import junit.framework.Test;
 import junit.framework.TestSuite;
